Fix catalog name synchronization issue
